Bug1629025 Server-Side Kyegen Enrollment (input/output)
This patch contains the code that attempts to provide the input and output
functionalities for Server-Side Keygen Enrollment.

It is currently not yet working.  It contains mainly the following pieces:

input:
  The new input plugin ServerKeygenP12PasswordInput.java, which is supposed
to work with the modification in ProfileSelect.template to accept the p12 passwd
that will be used to compose the p12 once the keys are generated on KRA and
cert issued by the CA.

output:
  The new output plugin PKCS12Output.java, which contains a test p12 to be sent
back to the browser, bypassing the actual keygen/issuance code (which is being
worked on separately) between the CA and KRA.

profile:
  The new caServerKeygen_UserCert.cfg profile which utilizes the new input and
output listed above.

On input side:
-------------------
What's expected:
  Once working, if you go to EE and click on (currently) the first profile:
"Manual User Dual-User Certificate Enrollment using server-side key generation",
one should expect to be able to specify the p12 password, p12 password
again (verified by the Javascript), the key type (RSA/ECC), key size/curve,
the subject name info, and the requestor info.
  Once filled out and submit, should be able to see the info reflected in the
debug log (add debug where needed).

What happends currently:
  Currently, even though I have a function passwdValidate() in ProfileSelect.template,
I have not yet found a proper place to put it. Also I don't think any of the info
filled out are passwd to the CA.
  Another thing is the selection for RSA/ECC is not showing.

On the output side:
-------------------
What's expected:
  Once working, if you add the pkcs12OutputImpl for output in a profile, it should
attempt to pass the test pkcs12 in the PKCS12Output.java code to the caller

What happens currently:
  I have not had a chance to find out where, once the p12 is created, the Http
 Response is called so the content type could be set to pkcs12 once the String
is converted back to bypte array with AtoB method.

Currently, I am using the following bug: Bug 1629025 - [RFE] Browser--based keygen is being deprecated, need replacement in pki;
However, it might be a good idea to create a server-side keygen specific bug, as this
is not intended to be a solution for all.  It's mainly for encryption keys, unless
the site administrator doesn't care about archiving signing keys.