Bug 1794213 Server-Side Kyegen Enrollment (input/output)
This patch contains the code that attempts to provide the input and output
functionalities for Server-Side Keygen Enrollment.
It includes test (hack) code to provide an end-to-end "demo".

The UI (Javascript) needs some work. Currently the keyType and keySize
selection part is where I need help with: ProfileSelect.template

This patch contains mainly the following pieces:

input:
  The new input plugin ServerKeygenInput.java, which is supposed
to work with the modification in ProfileSelect.template to
   - accept the p12 passwd that will be used to compose the p12 once the keys are generated on KRA and cert issued by the CA.
   - accept the keyType: RSA/ECC
   - accept the keySize: RSA key sizes or ECC curves

Profile:
  - The new default plugin: ServerKeygenUserKeyDefault.java, which contains test code to generate keys (instead of reaching out to KRA for now)
  - The new caServerKeygen_UserCert.cfg profile which utilizes the new input and output

output:
  The new output plugin PKCS12Output.java, which contains a test p12 to be sent
back to the browser, bypassing getting the actual p12.

What's expected:
  Once working, if you go to EE and click on (currently) the first profile:
"Manual User Dual-User Certificate Enrollment using server-side key generation",
one should expect to be able to specify the p12 password, p12 password
again (verified by the Javascript), the key type (RSA/ECC), key size/curve,
the subject name info, and the requestor info.
  Once filled out and submit, should be able to see the info reflected in the
debug log (add debug where needed).

What happends currently:
  Currently, once this patch applied, the CA instance should allow users to use
the new profile and submit a request (with the somewhat crippled keytype and keysize selection as described earlier).
  As an agent, one could approve the request.  While the cert is expected to be
issued, since the output is using the new pkcs12 output, it is not yet
displaying. That's because
  I have not had a chance to find out where, once the p12 is created, the Http
 Response is called so the content type could be set to pkcs12 once the String
is converted back to bypte array with AtoB method.

Finally:
Server-side key generation for enrollment is not intended to be a solution
for all.  It's mainly for encryption keys, unless the site administrator
 doesn't care about archiving signing keys.

https://bugzilla.redhat.com/show_bug.cgi?id=1794213