Fix 01-AddProfileCaAuditSigningCert
The upgrade script that is supposed to add caAuditSigningCert
profile into existing instances failed since the actual profile
configuration is packaged in pki-ca which is installed after
the upgrade is done.

To fix the problem, the upgrade script has been modified to
embed the content of the profile configuration.

https://bugzilla.redhat.com/show_bug.cgi?id=1883639
(cherry picked from commit 8533fdc51695d5965b7256ffbb73fd928097f7fd)