Bug1979710-TPS: separate config actions by profile permission list
This is just a patch to show where I think the permission screening
by allowed profiles could happen for TPS configurations,
using TokenService.java:retrieveTokensWithVLV() as an example.

If this looks like the right approach, then should add similar
screening to other retrieve/modify operations, and also for other areas
such as Activities and Certificates.

I wonder if there is a more efficient way of doing this, such as some
kind of a filter that says "if the tokneType field is A or B or C".

https://bugzilla.redhat.com/show_bug.cgi?id=1979710